requirements.txt
fastapi
uvicorn
python-multipart
pyjwt
numpy
Pillow
onnxruntime
opencv-python
insightface
